Add NAT44 1to1 SNAT and SDNAT support
The feature programs a list of nat4-1to1 actions.The match criteria is defined
as ACLs and attached to the interfaces. The ACLs are encoded with the value
that points to one of the nat44-1to1 actions. The feature checks for match in
both in2out and out2in directions and applies NAT on a match.